AEAD: Move `LessSafeKey` code into a submodule.
